Refactor resultproccesor
